Compute sum in convolution with double type so we dont loose precision in double datasets

+++ src/main/java/com/mycompany/imagej/ops/filter/convolve/MPIConvolveNaiveC.java
@@ -87,7 +87,7 @@ public class MPIConvolveNaiveC<I extends RealType<I>, O extends RealType<O> & Na
 					outC.localize(pos);
 
 					// kernel inlined version of the method convolve
-					float val = 0;
+					double val = 0;
 					inRA.setPosition(pos);
 
 					kernelC.reset();
